🎨 Style : interface명 변경
- interface명 변경
